Create object menu with icons.
