+++ mettascope/index.html
@@ -214,19 +214,36 @@
           </div>
         </div>
         <div class="actions">
-          <button class="mem-0">
-            <span class="amount">Mem=0</span>
-          </button>
-          <button class="mem-1">
-            <span class="amount">Mem=1</span>
-          </button>
-          <button class="mem-r">
-            <span class="amount">Mem=Rand</span>
+          <button class="memory">
+            <span class="text">Memory</span>
           </button>
         </div>
       </div>
       <img src="data/ui/panel-tip.png" alt="tip" class="tip">
       <img src="data/ui/close.png" alt="close.hover-icon" class="close hover-icon">
+    </div>
+    <div id="object-menu">
+      <div class="set-memory-to-0 hover">
+        <span class="text">Set Memory to 0</span>
+        <img src="data/ui/brain0.png" alt="icon" class="icon">
+      </div>
+      <div class="set-memory-to-1 hover">
+        <span class="text">Set Memory to 1</span>
+        <img src="data/ui/brain1.png" alt="icon" class="icon">
+      </div>
+      <div class="set-memory-to-random hover">
+        <span class="text">Scramble Memory</span>
+        <img src="data/ui/brainr.png" alt="icon" class="icon">
+      </div>
+      <div class="sep"></div>
+      <div class="copy-memory hover">
+        <span class="text">Copy Memory</span>
+        <img src="data/ui/copy.png" alt="icon" class="icon">
+      </div>
+      <div class="paste-memory hover">
+        <span class="text">Paste Memory</span>
+        <img src="data/ui/paste.png" alt="icon" class="icon">
+      </div>
     </div>
   </body>
 </html>
